refactor(test-sets): convert Garak import from Dialog to BaseDrawer
Replace the centered modal dialog with a right-anchored BaseDrawer
(720px wide) so Garak import is consistent with FileImportDrawer and
other test-set flows.

- Version chip and Preview button moved into the Select Probes header row
- Probe list Paper grows to fill available drawer height instead of being
  capped at a fixed max-height
- Footer wired via BaseDrawer props (cancel/close label, loading spinner,
  import button label logic preserved)
- State reset uses a useEffect on open instead of TransitionProps.onExited
- Rename GarakImportDialog → GarakImportDrawer (file, component, state var)
- Update e2e selectors from role=dialog to MuiDrawer-aware locator
